Don't immediately store new (and blank) entries
This adds behavior to allow for showing the entry details pane for a new entry
that will only be stored when the user clicks "save". Otherwise, they can
cancel out and nothing happens to the storage. This should make things behave
more nicely when we have this hooked up to the real storage backend.