Fix linting for real
The issue here was that npm apparently tries to be helpful when it sees globs
in the "scripts" section, only their implementation is wrong. If you specify
`src/**/*.js` and it finds a file like `src/*.js`, it doesn't look in any
subdirectories for more files! I have no idea why this is the case, but the
current code should work.