Update <ScrollingList/> selection on mousedown instead of click
This helps reduce (but not entirely eliminate) a temporary flash when you
click on an unselected item while the <ScrollingList/> is unfocused. When
mousedown happens, the list gets focus so the selected item changes color, and
then when the selection actually changes, it moves to the item you picked.

There is probably a better way to do this, but it should be ok for now.