(Attempt to) provide coverage for untested files
We're currently forced to load all our test and most of our src files into a
single omni-file. This is a mess; we're forced to exclude some files since
they'd break our tests, but then we lose coverage on them! Sadly, it's the
best we can get with the tooling we're using being in the (awful!) state that
it is.