Improve how the item list borders work
This implementation now uses ::before and ::after pseudo-elements to draw the
lines, allowing us to explicitly specify the z-index of each kind of border.
This ensures that the borders on the whole list go above normal items, but
*below* the borders on selected/focused items.

Note: there's one remaining bug here. There's a border between the search box
and the no-results placeholder when there shouldn't be. This will be addressed
in a later commit.